requirements.txt
google-generativeai
groq
python-dotenv
transformers
accelerate
torch
torchvision
torchaudio
transformers
scikit-learn
pandas
numpy
tqdm